Rename package from java to javaguide
This will both keeps better consistency with other sections
of our documentation but also avoid an annoying warning
that I’ve noticed while writing #7346. Here is an example
of this warning in our build:

https://travis-ci.org/playframework/playframework/jobs/232592636#L3259
